codeforces

+++ templates/number_theory/factorial_recursive.cpp
@@ -1,4 +1,4 @@
-int factorial (int n)
+unsigned long long factorial (unsigned long long n)
 {
     return (n == 0) || (n == 1) ? 1 : n * factorial(n-1);
 }
